Adjustments after code review
