+++ backend/package.json
@@ -15,8 +15,10 @@
     "@types/express": "^4.17.17",
     "@types/node": "^18.15.11",
     "cors": "^2.8.5",
+    "eslint-config-airbnb-base": "^15.0.0",
     "express-async-handler": "^1.2.0",
     "nodemon": "^2.0.22",
+    "prettier": "^2.8.7",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
   },
