+++ frontend/package.json
@@ -9,6 +9,8 @@
     "preview": "vite preview"
   },
   "devDependencies": {
+    "eslint-config-airbnb-base": "^15.0.0",
+    "prettier": "^2.8.7",
     "sass": "^1.62.0",
     "typescript": "^4.9.3",
     "vite": "^4.2.0"
